deploy: add README, fix typos
The csi-hostpath-snapshotter.yaml was misspelled when creating the
file. The upstream driver-registrar RBAC file uses
"csi-driver-registrar" as service account name and therefore the
hostpath plugin should also use that instead of the (now obsolete)
"csi-node-sa" name from the Kubernetes E2E testing.

This dependency on the upstream RBAC rules should be mentioned. For
the sake of avoiding yet another copy of URLs the new README just
refers to other documentation.