deploy: sidecar versions from .yaml
Picking the RBAC rules from a release branch is not quite accurate (we
want the rules released together with the tagged version) and is
something that must be kept in sync manually.

It is better to not hard-code this in the script and instead get the
exact versions from the .yaml files.

There should be no need for users to pass parameters to the script, so
let's just mention that this is possible without going into details in
the README.md. The previous instructions were out-dated, too.