deploy: allow overriding image registry and tags, use fixed images
They can be overridden for all images or per image by setting env
variables, for example:
  IMAGE_REGISTRY=localhost:9000 \
  IMAGE_TAG=canary \
  CSI_PROVISIONER_TAG=v1.0.2 \
     ./deploy/deploy.sh

This is useful for users in air-gapped clusters that host their images
in a local repository, for developers and for CI testing.

We switched to a release model where tags always refer to the same
image, therefore we don't need `imagePullPolicy: Always` anymore. We
also don't want it for CI testing when using `canary` images, because
then images can be loaded once onto a cluster before a test and are
guaranteed to not change in the middle of the test.

When using kind with locally build images, `imagePullPolicy: Always`
breaks deployment because the side-loaded image gets ignored.