deploy: update example for use with current images
The external-provisioner no longer accepts the --provisioner
parameter. While at it, parameters are now the same as in
external-attacher (same style, -v=5).

Right now it is still possible to use the same deployment for
Kubernetes 1.13 and Kubernetes master (aka "latest"). But soon we will
have to maintain two different deployments, one for Kubernetes 1.13
and one for Kubernetes 1.14. Moving the deployment script into a
sub-directory is a first step towards that.

The Kubernetes-CSI CI scripts expect to find directories or symlinks
to directories name "kubernetes-x.yy" and uses those when testing
against that Kubernetes version. If the same deployment works for
multiple Kubernetes releases, then the directory name is the oldest
supported Kubernetes release and the symlinks get added once it
becomes known that those also work.