deploy: unify and simplify deployment scripts
Conceptually, each deployment can have its own script. This may become
relevant at some point, so the CI looks for the script in the
individual directories.

But in practice the current two copies of the script were
identical. It makes sense to keep only a single copy in a neutral
place and use it via symlinks.

While at it, the unused CRD installation gets removed. It's definitely
not needed for 1.14 (no CRDs) and should not be needed for
1.13 (cluster setup must install them).