per-volume ephemeral mode
This removes the limitation that the driver can only be deployed for
one mode or the other. Instead the driver uses the new
csi.storage.k8s.io/ephemeral
field (https://github.com/kubernetes/kubernetes/issues/79624) in the
volume context to determine how it should behave in NodePublishVolume.

Supporting both modes in the same deployment makes it possible to test
ephemeral mode without having to change how the Prow jobs deploy the
driver, which is once as part of the cluster setup.

The --ephemeral parameter is still supported and can be used to force
the driver into ephemeral mode, which is useful when deploying on
Kubernetes 1.15 where csi.storage.k8s.io/ephemeral is not supported.