deploy: fix canary jobs on 1.14/15/16
The Prow canary jobs correctly showed that the upcoming
external-snapshotter 2.0.0 is not going to be compatible with the
existing deployments because of API changes. But now that this has
been discovered, it's not useful anymore to keep those jobs in the
failing state.

We need them to run again until some other sidecar breaks its
API. This can be achieved by documenting that external-snapshotter
cannot be updated to its canary version and then just keeping it at
the version specified for the deployment. The same can be done for
other sidecars if the need arises.