deploy: move test driver definition into driver repo
This will enable the removal of hostpath specific code from
csi-release-tools. The deploy script (now using the name "deploy.sh"
because that can be the same for all CSI drivers) is asked to create a
test driver configuration in addition to installing the driver.

The script could do that dynamically. But in the CSI driver hostpath
repo we simply keep one test-driver.yaml per deployment, which works
based on the assumption that the E2E test suite that corresponds to
the Kubernetes version is used for testing, which is the case
nowadays.

These three test-driver.yaml are indeed different because the name of
capabilities changed over time. prow.sh used to have a union of all
those names. Now each file really only uses the name that is
applicable.

"topology" support hadn't been enabled in prow.sh. This looks like an
oversight, so it gets added now.