update pkg/volume/util/volumepathhandler to 1.20.0
The RemoveLoopDevice was made internal and replaced with the DetachFileDevice
wrapper function which already includes code to look up the device:

func (v VolumePathHandler) DetachFileDevice(path string) error {
       loopPath, err := v.GetLoopDevice(path)
       if err != nil {
               if err.Error() == ErrDeviceNotFound {
                       klog.Warningf("DetachFileDevice: Couldn't find loopback device which takes file descriptor lock. device path: %q", path)
               } else {
                       return err
               }
       } else {
               if len(loopPath) != 0 {
                       err = removeLoopDevice(loopPath)
                       if err != nil {
                               return err
                       }
               }
       }
       return nil
}

The advantage is that the newer function continues when the device is
not found, which is what we want when cleaning up in an idempotent
way.

That it logs a warning in that case is less nice.