deploy: update images, add Kubernetes 1.20
Kubernetes 1.17 is no longer supported and by updating to 1.18 as
minimal version we can use the GA CSIDriver v1 API.

Also, all sidecar and driver images get updated. The CSI hostpath
image update is important because the recently added health check
sidecar depends on the new version.

Kubernetes 1.20 is where officially the v1 snapshotter API is
introduced. Therefore the new deployment for it also uses the
external-snapshotter 4.x and the v1 APIs.