support for capacity tracking + distributed provisioning
This fakes capacity by pretending to have linear storage for different
kinds and subtracting the size of existing volumes from that.

A test deployment with some example storage classes, an example app
with a generic ephemeral inline volume, and storage capacity tracking
enabled is provided for use on a Kubernetes cluster where these alpha
features are enabled.

When that feature is not enabled in the cluster, also the driver
deployment is done without storage capacity enabled. This then serves
as a test that distributed provisioning works.

prow.sh can be used to test this new deployment, for example with:

CSI_PROW_SANITY_POD=csi-hostpath-socat-0 \
  CSI_PROW_SANITY_CONTAINER=socat \
  CSI_PROW_DEPLOYMENT=kubernetes-distributed \
  CSI_PROW_KUBERNETES_VERSION=1.19.0 \
  ./.prow.sh