CSI ephemeral volumes: use fixed size
With storage capacity simulation enabled, trying to create ephemeral
volumes with the default maximum volume size may have failed:

  Warning  FailedMount  6s (x5 over 14s)  kubelet            MountVolume.SetUp failed for volume "ephemeral-volume" : rpc error: code = ResourceExhausted desc = not enough capacity: have 100Gi, need 1Ti

Using a smaller size is better because then more than one volume can
be created per host. Even better would be to also add a parameter for
it, but that can be added later.