deploy: fix CSIStorageCapacity check on alpha clusters
"kubectl api-resources" does not list v1alpha1.storage.k8s.io even
when it is enabled, presumably because it gets superseeded by
v1beta1.storage.k8s.io. That means we need to go back to trying to use
the resource and grepping the output. To get around "-o pipefail", the
return code is explicitly ignored inside the pipe.