deploy: better diagnostics when driver deployment fails
With the recently introduced labels it becomes possible to describe
all objects that make up the driver deployment. This might be a lot of
output, but as this typically runs in a CI environment, it is better
to print too much than to little information because retrieving more
information later isn't possible.

Now also the logs of all containers are printed.