relax volume lifecycle checks by default
The recently introduced "still in use" check revealed a bug in
Kubernetes (https://github.com/kubernetes/kubernetes/issues/101911). While
the check itself is correct, enabling it would cause a lot of test
flakes. Therefore the check has to be disabled until all Kubernetes
versions that we test against are fixed.