NodePublishVolume: require parents of target directory to exist
According to the spec, "the CO SHALL ensure that the parent directory
of this path exists". The driver didn't rely on that and created those
parent directories itself if necessary, without removing them in
NodeUnpublishVolume.

This hides potential bugs in the CO or incorrect configuration of
csi-sanity (like creating directories in the wrong pod). Therefore it
is better to be strict and only create the actual target directory
itself.