ensure that volumes get rescheduled when fake capacity is exhausted
The driver must return ResourceExhausted when it doesn't have capacity, not
OutOfRange. Then external-provisioner knows that it needs to cause
rescheduling. Also, that error must not get wrapped because that hides
the status code.

The incorrect code bug was introduced in 6ad318936 via cut-and-paste. For the
case above where the volume is too large for *all* driver instances because it
exceeds the configured maximum volume size it makes sense to return OutOfRange
because the error is permanent.