Add `namespace: default` to Service and StatefulSets
When the driver is deployed while the current context is set to
something else than `default`, things go wrong as the Service and
StatefulSets are created in the current namespace.

By adding the namespace explicitly in the Service and StatefulSets, all
components of the driver get deployed correctly in the namespace called
`default`.