Allow ignoring unreadable files when snapshotting volumes
Snapshots are implemented using `tar` for file system volumes. With this
patch, users can set the `ignoreFailedRead` parameter in the
VolumeSnapshotClass parameter to `true`, which results in the
`--ignore-failed-read` option to be passed to `tar`.

This enhancement is particularly useful for end-to-end testing of
applications that may delete files initially included when the tar
process started and retrieved the file list. By ignoring unreadable
files, the snapshot process can finish successfully.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>