Add --enable-list-snapshots command-line option
This commit introduces the --enable-list-snapshots command-line option,
allowing users to explicitly enable or disable the
ControllerServiceCapability_RPC_LIST_SNAPSHOTS capability.

By default, this option is set to true to preserve existing behavior.

Disabling RPC_LIST_SNAPSHOTS is useful for emulating the behavior of CSI
drivers that do not support this optional RPC call.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>